+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## Changes
 - Removed the relationship event attribute key prefix (#300)
 - Removed the user_block event attribute key prefix (#291)
-- Changed posts' optional data representation (#272) 
+- Changed posts' optional data representation (#272)
+- Implemented the possibility to refuse and cancel DTag requests from both receiver and sender (#297)
 
 ## Bug fixes
 - Fixed the possibility of requesting a transfer of an empty DTag (#292)